Refactor staff shift management UI and logic
Moved staff shift management templates from 'shift' to 'staff_shift' directory and updated controller/service to use StaffShiftDTO instead of entity. Adjusted controller endpoints and service methods for improved DTO usage and streamlined form handling. Updated StaffShiftDTO to use Integer IDs. Removed obsolete templates and added new ones for staff shift management.

+++ PizzarIO_SWP391/src/main/java/com/group1/swp/pizzario_swp391/dto/staffshift/StaffShiftDTO.java
@@ -12,9 +12,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class StaffShiftDTO {
-    private Long staffId;
-    private Long shiftId;
-
+    private Integer staffId;
+    private Integer shiftId;
     // Ngày làm, trạng thái ca làm, lương/giờ
     private LocalDate workDate;
     private String status; // ví dụ: SCHEDULED, CHECKED_IN, CHECKED_OUT
